+++ src/server/pps/server/s3g_sidecar_proxy.go
@@ -1,0 +1,143 @@
+// A proxy to the real S3 server underlying pachyderm, for cases when stronger
+// S3 compatibility & performance is required for "s3_out" feature.
+
+package server
+
+import (
+	"fmt"
+	"log"
+	"net/http"
+	"net/http/httputil"
+	"net/url"
+	"os"
+	"strings"
+	"time"
+
+	"github.com/gorilla/mux"
+	"github.com/sirupsen/logrus"
+)
+
+type RawS3Proxy struct {
+}
+
+func (r *RawS3Proxy) ListenAndServe(port uint16) error {
+	// os.GetEnv("STORAGE_BACKEND") ==> "MINIO" or, presumably, "AWS"...
+	// MINIO_SECRET, MINIO_ID, MINIO_ENDPOINT (e.g.
+	// minio.default.svc.cluster.local:9000), MINIO_SECURE, MINIO_BUCKET also
+	// set.
+
+	// TODO!!! Need to mutate the bucket name in the proxied requests from
+	// "out" to whatever the real backend bucket is called in the backend...
+	// Before we do this, we'll just modify the user code to use the real
+	// minio bucket name, just to prove the concept.
+
+	if os.Getenv("STORAGE_BACKEND") != "MINIO" {
+		panic("only minio supported by proxy to real backend s3_out feature right now - TODO: add real AWS!")
+	}
+
+	proxyRouter := mux.NewRouter()
+	proxyRouter.PathPrefix("/").HandlerFunc(
+		func(w http.ResponseWriter, r *http.Request) {
+
+			logrus.Infof("===> IN PROXY ROUTER, w=%+v, r=%+v", w, r)
+
+			proxy := &httputil.ReverseProxy{
+				Director: func(req *http.Request) {
+					u := os.Getenv("MINIO_ENDPOINT")
+					// TODO: check MINIO_SECURE
+					// TODO: use MINIO_SECRET, MINIO_ID, ... need to sign
+					// (after unwrapping??) the request before we send it
+					// onwards: see https://github.com/smarty-archives/go-aws-auth (deprecated!) etc
+					// TODO: need to add r.URL.Path?
+					logrus.Infof("PROXY: r.URL.Path=%s", r.URL.Path)
+					target, err := url.Parse(fmt.Sprintf("http://%s", u))
+					if err != nil {
+						log.Fatal(err)
+					}
+					targetQuery := target.RawQuery
+					req.URL.Scheme = target.Scheme
+					req.URL.Host = target.Host
+					req.URL.Path, req.URL.RawPath = joinURLPath(target, req.URL)
+					if targetQuery == "" || req.URL.RawQuery == "" {
+						req.URL.RawQuery = targetQuery + req.URL.RawQuery
+					} else {
+						req.URL.RawQuery = targetQuery + "&" + req.URL.RawQuery
+					}
+					if _, ok := req.Header["User-Agent"]; !ok {
+						// explicitly disable User-Agent so it's not set to default value ???
+						req.Header.Set("User-Agent", "")
+					}
+					// // If calling the istio ingress, we need to set the endpoint host in the header
+					// if coords.Host != "" {
+					// 	// req.Header.Set("Host", coords.Host)
+					// 	req.Host = coords.Host
+					// }
+				},
+				ErrorHandler: func(rw http.ResponseWriter, r *http.Request, err error) {
+					log.Printf("Error from proxy connection for %s %s: %s", r.Host, r.URL.Path, err)
+					http.Error(w, err.Error(), http.StatusInternalServerError)
+				},
+			}
+
+			proxy.ServeHTTP(w, r)
+		},
+	)
+	// log where we're listening
+	logrus.Infof("PROXY: Listening on port %d", port)
+	srv := &http.Server{
+		Handler: proxyRouter,
+		Addr:    fmt.Sprintf("0.0.0.0:%d", port),
+		// Good practice: enforce timeouts for servers you create!
+		WriteTimeout: 15 * time.Second,
+		ReadTimeout:  15 * time.Second,
+	}
+	return srv.ListenAndServe()
+}
+
+// TODO: factor this out into a separate file, at least...
+// from https://gist.github.com/thezelus/d5ac9ec563b061c514dc ...
+// func handler(target string, p *httputil.ReverseProxy) func(http.ResponseWriter, *http.Request) {
+// 	return func(w http.ResponseWriter, r *http.Request) {
+// 		r.URL.Host = url.Host
+// 		r.URL.Scheme = url.Scheme
+// 		r.Header.Set("X-Forwarded-Host", r.Header.Get("Host"))
+// 		r.Host = url.Host
+
+// 		r.URL.Path = mux.Vars(r)["rest"]
+// 		p.ServeHTTP(w, r)
+// 	}
+// }
+
+// tf we need all this cruft
+func singleJoiningSlash(a, b string) string {
+	aslash := strings.HasSuffix(a, "/")
+	bslash := strings.HasPrefix(b, "/")
+	switch {
+	case aslash && bslash:
+		return a + b[1:]
+	case !aslash && !bslash:
+		return a + "/" + b
+	}
+	return a + b
+}
+
+func joinURLPath(a, b *url.URL) (path, rawpath string) {
+	if a.RawPath == "" && b.RawPath == "" {
+		return singleJoiningSlash(a.Path, b.Path), ""
+	}
+	// Same as singleJoiningSlash, but uses EscapedPath to determine
+	// whether a slash should be added
+	apath := a.EscapedPath()
+	bpath := b.EscapedPath()
+
+	aslash := strings.HasSuffix(apath, "/")
+	bslash := strings.HasPrefix(bpath, "/")
+
+	switch {
+	case aslash && bslash:
+		return a.Path + b.Path[1:], apath + bpath[1:]
+	case !aslash && !bslash:
+		return a.Path + "/" + b.Path, apath + "/" + bpath
+	}
+	return a.Path + b.Path, apath + bpath
+}
